backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==5.0.0
google-generativeai==0.3.2
pandas
openpyxl